our_solution/backend/requirements.txt
openai>=1.0.0
Flask>=3.0.0
Flask-Cors>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pytz>=2024.1
